C4267の警告エラーを解決

+++ Enemy.cpp
@@ -60,9 +60,9 @@ void Enemy::Initialize(Model* model, Model* modelFace, uint32_t textureHandle)
 		getline(enemyfile, stringweakPower);
 		getline(enemyfile, stringheavyPower);
 
-		int pos1 = stringhp.find(":");
-		int pos2 = stringweakPower.find(":");
-		int pos3 = stringheavyPower.find(":");
+		int pos1 = static_cast<int> (stringhp.find(":"));
+		int pos2 = static_cast<int>	(stringweakPower.find(":"));
+		int pos3 = static_cast<int>	(stringheavyPower.find(":"));
 
 		if (pos1 != string::npos) {
 			stringhp = stringhp.substr(pos1 + 1);
